requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.17.0
numpy==1.26.4
Pillow==10.4.0
Jinja2==3.1.4
python-multipart==0.0.10
pydantic==2.9.2
starlette==0.38.6
anyio==4.6.0
h11==0.14.0